update shopify implementation
